+++ VirtualShelter/src/main/java/hu/bme/aut/vshelter/entity/Institution.java
@@ -71,7 +71,7 @@ public class Institution extends Advertiser {
 	public int hashCode() {
 		final int prime = 31;
 		int result = 1;
-		result = prime * result + id;
+		result = prime * result + (int)id;
 		return result;
 	}
